+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/filter/controller/response/AssigneeFilterResponse.java
@@ -1,6 +1,6 @@
 package codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.controller.response;
 
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.service.information.AssigneeFilterInformation;
+import codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.service.information.AssigneeFilterInformation;
 import lombok.Builder;
 import lombok.Getter;
