+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/filter/controller/response/AuthorFilterResponse.java
@@ -1,6 +1,6 @@
 package codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.controller.response;
 
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.service.information.AuthorFilterInformation;
+import codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.service.information.AuthorFilterInformation;
 import lombok.Builder;
 import lombok.Getter;
