+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/filter/controller/response/MilestoneFilterResponse.java
@@ -1,6 +1,6 @@
 package codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.controller.response;
 
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.service.information.MilestoneFilterInformation;
+import codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.service.information.MilestoneFilterInformation;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import lombok.Builder;
