+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/filter/service/FilterService.java
@@ -1,6 +1,6 @@
-package codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.service;
+package codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.service;
 
-import codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.service.information.FilterListInformation;
+import codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.service.information.FilterListInformation;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.label.repository.LabelRepository;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.label.repository.VO.LabelDetailsVO;
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.member.repository.MemberRepository;
