+++ be/src/main/java/codesquad/kr/gyeonggidoidle/issuetracker/domain/filter/service/information/AssigneeFilterInformation.java
@@ -1,4 +1,4 @@
-package codesquad.kr.gyeonggidoidle.issuetracker.domain.issue.service.information;
+package codesquad.kr.gyeonggidoidle.issuetracker.domain.filter.service.information;
 
 import codesquad.kr.gyeonggidoidle.issuetracker.domain.member.repository.vo.MemberDetailsVO;
 import lombok.Builder;
